+++ experiments/ClimaEarth/run_dry_held_suarez.jl
@@ -156,7 +156,7 @@ end
 =#
 
 ## coupler exchange fields
-coupler_field_names = (
+coupler_field_names = [
     :T_sfc,
     :z0m_sfc,
     :z0b_sfc,
@@ -176,33 +176,32 @@ coupler_field_names = (
     :P_net,
     :temp1,
     :temp2,
-)
-coupler_fields =
-    NamedTuple{coupler_field_names}(ntuple(i -> ClimaCore.Fields.zeros(boundary_space), length(coupler_field_names)))
+]
+coupler_fields = Interfacer.init_coupler_fields(FT, coupler_field_names, boundary_space)
 
 ## model simulations
 model_sims = (atmos_sim = atmos_sim,);
 
-## dates
-date0 = date = Dates.DateTime(start_date, Dates.dateformat"yyyymmdd")
-dates = (; date = [date], date0 = [date0], date1 = [Dates.firstdayofmonth(date0)], new_month = [false])
+## start date
+start_date = Dates.DateTime(start_date, Dates.dateformat"yyyymmdd")
 
 #=
 ## Initialize Callbacks
 =#
-schedule_checkpoint = EveryCalendarDtSchedule(TimeManager.time_to_period(checkpoint_dt); start_date = date0)
+schedule_checkpoint = EveryCalendarDtSchedule(TimeManager.time_to_period(checkpoint_dt); start_date)
 checkpoint_cb = TimeManager.Callback(schedule_checkpoint, Checkpointer.checkpoint_sims)
 
 callbacks = (; checkpoint = checkpoint_cb)
 
 cs = Interfacer.CoupledSimulation{FT}(
     comms_ctx,
-    dates,
+    Ref(start_date),
     boundary_space,
     coupler_fields,
     nothing, # conservation checks
     [tspan[1], tspan[2]],
     Δt_cpl,
+    Ref(tspan[1]),
     model_sims,
     callbacks,
     dir_paths,
@@ -235,8 +234,8 @@ function solve_coupler!(cs)
 
     ## step in time
     walltime = @elapsed for t in ((tspan[begin] + Δt_cpl):Δt_cpl:tspan[end])
-        # Update date
-        cs.dates.date[] = TimeManager.current_date(cs, t)
+        # Update current time
+        cs.t[] = t
 
         ## step sims
         FieldExchanger.step_model_sims!(cs.model_sims, t)
@@ -244,7 +243,7 @@ function solve_coupler!(cs)
         FieldExchanger.import_atmos_fields!(cs.fields, cs.model_sims, cs.boundary_space, cs.turbulent_fluxes) # radiative and/or turbulent
 
         ## callback to checkpoint model state
-        TimeManager.maybe_trigger_callback(cs.callbacks.checkpoint, cs, t)
+        TimeManager.maybe_trigger_callback(cs.callbacks.checkpoint, cs)
     end
     @info(walltime)